Correction et simplifications
