+++ src/main/java/com/ipiecoles/java/java220/Technicien.java
@@ -74,6 +74,6 @@ public class Technicien extends Employe implements Comparable<Technicien> {
 
 	@Override
 	public int compareTo(Technicien o) {
-		return Integer.compare(o.getGrade(), this.grade);
+		return Integer.compare(this.grade, o.getGrade());
 	}
 }
